Moved out phonon and harmonic functions from quasi_harmonic.py to phonons.py
